+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/TestFrameworkType.kt
@@ -6,26 +6,20 @@ import org.jetbrains.intellij.platform.gradle.models.Coordinates
 
 /**
  * Definition of Test Framework types available for writing tests for IntelliJ Platform plugins.
+ *
+ * @param coordinates Maven coordinates of test framework artifact.
  */
-interface TestFrameworkType {
+sealed class TestFrameworkType(val coordinates: Coordinates) {
+    object Platform : TestFrameworkType(Coordinates("com.jetbrains.intellij.platform", "test-framework"))
+    object JUnit5 : TestFrameworkType(Coordinates("com.jetbrains.intellij.platform", "test-framework-junit5"))
+    object Bundled : TestFrameworkType(Coordinates("bundled", "lib/testFramework.jar"))
 
-    /**
-     * Maven coordinates of test framework artifact;
-     */
-    val coordinates: Coordinates;
-
-    enum class Platform(override val coordinates: Coordinates) : TestFrameworkType {
-        JUnit4(Coordinates("com.jetbrains.intellij.platform", "test-framework")),
-        JUnit5(Coordinates("com.jetbrains.intellij.platform", "test-framework-junit5")),
-        Bundled(Coordinates("bundled", "lib/testFramework.jar")),
-    }
-
-    enum class Plugin(override val coordinates: Coordinates) : TestFrameworkType {
-        Go(Coordinates("com.jetbrains.intellij.go", "go-test-framework")),
-        Ruby(Coordinates("com.jetbrains.intellij.idea", "ruby-test-framework")),
-        Java(Coordinates("com.jetbrains.intellij.java", "java-test-framework")),
-        JavaScript(Coordinates("com.jetbrains.intellij.javascript", "javascript-test-framework")),
-        Maven(Coordinates("com.jetbrains.intellij.maven", "maven-test-framework")),
-        ReSharper(Coordinates("com.jetbrains.intellij.resharper", "resharper-test-framework")),
+    object Plugin {
+        object Go : TestFrameworkType(Coordinates("com.jetbrains.intellij.go", "go-test-framework"))
+        object Ruby : TestFrameworkType(Coordinates("com.jetbrains.intellij.idea", "ruby-test-framework"))
+        object Java : TestFrameworkType(Coordinates("com.jetbrains.intellij.java", "java-test-framework"))
+        object JavaScript : TestFrameworkType(Coordinates("com.jetbrains.intellij.javascript", "javascript-test-framework"))
+        object Maven : TestFrameworkType(Coordinates("com.jetbrains.intellij.maven", "maven-test-framework"))
+        object ReSharper : TestFrameworkType(Coordinates("com.jetbrains.intellij.resharper", "resharper-test-framework"))
     }
 }